Task2: move route handlers to the 'user' folder
